refactor: split stacked_video.rs into directory along feature gates
Convert the 1,560-line stacked_video.rs into a directory module:

- stacked_video/mod.rs (611 lines): GridLayout, pack/unpack, tests
- stacked_video/encoder.rs (208 lines): StackedEncoder + config
- stacked_video/replay.rs (559 lines): ReplayRecordingSource,
  SessionStackedRecorder, GpuAtlasRecorder
- stacked_video/source.rs (163 lines): StackedSource (read-side)

Sub-modules gated on #[cfg(feature = "stacked-output")] matching the
original feature boundaries. No re-exports.

All tests pass, clippy clean.
